Widget view:widget, UI fixes, db update, db delete, ets added

+++ app/src/main/java/com/tanka/accessories/todoroom/views/NotesAdapter.java
@@ -1,16 +1,25 @@
 package com.tanka.accessories.todoroom.views;
 
+import android.content.DialogInterface;
+import android.content.Intent;
+import android.graphics.drawable.AnimationDrawable;
+import android.os.Parcelable;
+import android.support.v7.app.AlertDialog;
 import android.support.v7.widget.RecyclerView;
 import android.util.Log;
+import android.view.GestureDetector;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.ImageView;
+import android.widget.LinearLayout;
 import android.widget.TextView;
 import android.widget.Toast;
 
 import com.tanka.accessories.todoroom.R;
 import com.tanka.accessories.todoroom.data.model.Note;
 
+import java.io.Serializable;
 import java.util.List;
 
 /**
@@ -21,6 +30,8 @@ public class NotesAdapter extends RecyclerView.Adapter<NotesAdapter.ViewHolder> 
 
     private List<Note> itemList;
     MainActivity activity;
+    private static final long DOUBLE_CLICK_TIME_DELTA = 300;
+    private long lastClickTime = 0;
 
     public NotesAdapter(MainActivity activity, List<Note> itemList) {
         this.itemList = itemList;
@@ -46,13 +57,48 @@ public class NotesAdapter extends RecyclerView.Adapter<NotesAdapter.ViewHolder> 
         holder.body.setText(itemList.get(listPosition).getBody());
         holder.type.setText(itemList.get(listPosition).getType());
 
+        AnimationDrawable animationDrawable = (AnimationDrawable) holder.contentLayout.getBackground();
+        if (listPosition % 2 == 0) {
+            animationDrawable.setEnterFadeDuration(2500);
+            animationDrawable.setExitFadeDuration(2500);
+        } else {
+            animationDrawable.setEnterFadeDuration(5000);
+            animationDrawable.setExitFadeDuration(10000);
+        }
+
+        animationDrawable.start();
+
         holder.itemView.setOnClickListener(v -> {
-            Toast.makeText(activity,itemList.get(listPosition).getTitle(),Toast.LENGTH_LONG).show();
+
+            Note note = itemList.get(listPosition);
+            Intent intent = new Intent(activity, NoteDetail.class);
+            intent.putExtra("note", note);
+            activity.startActivity(intent);
+
         });
         holder.itemView.setOnLongClickListener(v -> {
-            deleteItem(itemList.get(listPosition));
+            showDeleteDialog(listPosition);
             return true;
         });
+
+        holder.ivEdit.setOnClickListener(v -> {
+            activity.editNote(itemList.get(listPosition));
+        });
+    }
+
+    private void showDeleteDialog(int listPosition) {
+        AlertDialog.Builder builder = new AlertDialog.Builder(activity);
+        builder.setTitle(R.string.text_delete);
+        builder.setMessage(R.string.info_delete);
+        builder.setPositiveButton(R.string.yes, (dialog, which) -> {
+
+            deleteItem(itemList.get(listPosition));
+            dialog.dismiss();
+        });
+        builder.setNegativeButton(R.string.no, (dialog, which) -> {
+            dialog.dismiss();
+        });
+        builder.show();
     }
 
     static class ViewHolder extends RecyclerView.ViewHolder implements View.OnClickListener {
@@ -60,6 +106,8 @@ public class NotesAdapter extends RecyclerView.Adapter<NotesAdapter.ViewHolder> 
         public TextView date;
         public TextView body;
         public TextView type;
+        public ImageView ivEdit;
+        private LinearLayout contentLayout;
 
         public ViewHolder(View itemView) {
             super(itemView);
@@ -67,7 +115,10 @@ public class NotesAdapter extends RecyclerView.Adapter<NotesAdapter.ViewHolder> 
             title = itemView.findViewById(R.id.tvTitle);
             date = itemView.findViewById(R.id.tvDate);
             body = itemView.findViewById(R.id.tvBody);
+            ivEdit = itemView.findViewById(R.id.ivEdit);
             type = itemView.findViewById(R.id.tvType);
+            contentLayout = itemView.findViewById(R.id.contentLayout);
+
         }
 
         @Override
@@ -77,6 +128,7 @@ public class NotesAdapter extends RecyclerView.Adapter<NotesAdapter.ViewHolder> 
 
 
     }
+
     private void deleteItem(Note item) {
         activity.deleteNote(item);
